[修改]1. 修改消息操作类型为枚举

+++ GameFrameX.Client/UnityTcpClient.cs
@@ -132,7 +132,7 @@ public class UnityTcpClient
 
         MessageObjectHeader messageObjectHeader = new MessageObjectHeader
         {
-            OperationType = (byte)message.OperationType,
+            OperationType = message.OperationType,
             UniqueId = message.UniqueId,
             MessageId = message.MessageId,
         };
